+++ doc/changelog.rst
@@ -3,6 +3,7 @@
 Release Notes
 =============
 
+- :bug:`-` When updating the settings for an event with a custom domain, pretalx would mistakenly show an error message that this domain was already in use for another event.
 - :release:`1.1.2 <2020-01-10>`
 - :bug:`-` If a CfP requested the submitters' availability at submission time, only logged-in users could successfully finish the submission workflow.
 - :release:`1.1.1 <2020-01-10>`